requirements.txt
backports.csv==1.0.7
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.7
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.2.0
django-reversion==3.0.3
djangorestframework==3.9.2
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.9.2
idna==2.8
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
MarkupSafe==1.1.1
odfpy==1.4.0
openpyxl==2.6.1
Pillow==5.4.1
Pygments==2.3.1
pytz==2018.9
PyYAML==5.1
requests==2.21.0
six==1.12.0
tablib==0.13.0
uritemplate==3.0.0
urllib3==1.24.1
wincertstore==0.2
xadmin2==2.0.1
xlrd==1.2.0
XlsxWriter==1.1.5
xlwt==1.3.0
